+++ RespiraAMS-Platform/RespiraAMS-Platform.Doctor/Infrastructure/DependencyInjection.cs
@@ -30,7 +30,7 @@ namespace Infrastructure
             builder.Services.AddMappingProfiles();
         }
 
-        public static void ApplyMigrations(this IHost host)
+        public static void ApplyMigrations(this IHost host, bool isDevEnv)
         {
             using var scope = host.Services.CreateScope();
             var context = scope.ServiceProvider.GetRequiredService<DoctorDbContext>();
@@ -40,7 +40,7 @@ namespace Infrastructure
             }
             catch (Exception)
             {
-                context.Database.EnsureDeleted();
+                if (isDevEnv) context.Database.EnsureDeleted();
                 context.Database.Migrate();
             }
         }
